P401 auto render

See merge request students/A1_Maze.ID_356285/Team__TL__myregree_student.21_school.ru_.fQ1slrnxSme5_qjWhO8UvQ/A1_Maze-1!24

+++ src/views/maze_canvas.h
@@ -21,8 +21,8 @@ class MazeCanvas : public QWidget {
   s21::Maze *maze_;
   int cols_;
   int rows_;
-  int cell_width_ = 0;
-  int cell_height_ = 0;
+  double cell_width_ = 0;
+  double cell_height_ = 0;
   int click_counter_ = 0;
   std::vector<std::pair<int, int>> cells_;
